Adding GraphQL and Server Refactor
