+++ graph/resolvers.js
@@ -1,0 +1,5 @@
+module.exports = {
+  Query: {
+    hello: () => 'world',
+  }
+};
