+++ graph/typeDefs.graphql
@@ -1,0 +1,7 @@
+type Query {
+  hello: String
+}
+
+schema {
+  query: Query
+}
