+++ package.json
@@ -29,6 +29,9 @@
     "file-loader": "0.11.2",
     "fs-extra": "3.0.1",
     "googleapis": "^22.2.0",
+    "graphql": "^0.11.7",
+    "graphql-server-express": "^1.2.0",
+    "graphql-tools": "^2.7.2",
     "html-webpack-plugin": "2.29.0",
     "jest": "20.0.4",
     "object-assign": "4.1.1",
