+++ routes/index.js
@@ -1,0 +1,23 @@
+const express = require('express');
+
+const router = express.Router();
+
+// =============================================================================
+// STATICS
+// =============================================================================
+
+router.use(express.static('build'));
+
+// =============================================================================
+// GRAPHQL ROUTER
+// =============================================================================
+
+router.use('/graph', require('./graph'));
+
+// =============================================================================
+// ROUTES
+// =============================================================================
+
+router.use('/api', require('./api'));
+
+module.exports = router;
