+++ server.js
@@ -1,15 +1,14 @@
 const express = require('express');
-const google = require('googleapis');
 const bodyParser = require('body-parser');
-const axios = require('axios');
-
-const credentials = require('./creds.js');
+const routes = require('./routes');
 
 const app = express();
 
-app.use(express.static('build'));
+// =============================================================================
+// APPLICATION MIDDLEWARE
+// =============================================================================
 
-app.use(bodyParser.urlencoded({ extended: true }));
+app.use(bodyParser.urlencoded({extended: true}));
 app.use(bodyParser.json());
 
 app.use((req, res, next) => {
@@ -18,100 +17,10 @@ app.use((req, res, next) => {
   next();
 });
 
-const auth = new google.auth.JWT(
-  credentials.client_email,
-  null,
-  credentials.private_key,
-  ['https://www.googleapis.com/auth/spreadsheets'],
-  null
-);
+// =============================================================================
+// ROUTES
+// =============================================================================
 
-google.options({auth});
-
-const sheets = google.sheets('v4');
-const spreadsheetId = '18Q3kTrNtTYUyscylEly5mMms_n9g_sj0IPdAnn-9EME';
-
-const getMoviesIds = (req, res) => {
-  return new Promise((resolve, reject) => {
-    sheets.spreadsheets.values.get({
-      spreadsheetId,
-      range: 'diversidadmedia!B7:B'
-    }, (err, response) => {
-      resolve(response.values[0]);
-    });
-  });
-}
-
-const request = axios.create({
-  baseURL: 'https://api.themoviedb.org'
-});
-
-request.defaults.headers.common['Authorization'] = `Bearer ${credentials.tmdb.access_token}`;
-request.defaults.headers.common['Content-Type'] = `application/json;charset=utf-8`;
-
-app.get('/update', async (req, res, next) => {
-  try {
-    const ids = await getMoviesIds();
-
-    ids.map(async (id) => {
-      try {
-        await request.post(`/3/list/${credentials.tmdb.list}/add_item?api_key=${credentials.tmdb.api_key}`, {
-          media_id: id
-        }); 
-      } catch (err) {
-        return next(err);
-      }
-    })
-
-    res.sendStatus(200);
-  } catch (err) {
-    next(err);
-  }
-})
-
-app.get('/movies', async ({query: {page}}, res, next) => {
-  try {
-    const {data} = await request.get(`/4/list/${credentials.tmdb.lists.movies}?api_key=${credentials.tmdb.api_key}&page=${page}`);
-    res.send(data);
-  } catch (err) {
-    next(err);
-  }
-});
-
-app.get('/shows', async ({query: {page}}, res, next) => {
-  try {
-    const {data} = await request.get(`/4/list/${credentials.tmdb.lists.shows}?api_key=${credentials.tmdb.api_key}&page=${page}`);
-    res.send(data);
-  } catch (err) {
-    next(err);
-  }
-});
-
-app.get('/list/:id', async ({params: {id}}, res, next) => {
-  try {
-    const {data} = await request.get(`/4/list/${id}?api_key=${credentials.tmdb.api_key}`);
-    res.send(data);
-  } catch (err) {
-    next(err);
-  }
-});
-
-app.get('/movie/:id', async ({params: {id}}, res, next) => {
-  try {
-    const {data} = await request.get(`/3/movie/${id}?api_key=${credentials.tmdb.api_key}`);
-    res.send(data);
-  } catch (err) {
-    next(err);
-  }
-});
-
-app.get('/show/:id', async ({params: {id}}, res, next) => {
-  try {
-    const {data} = await request.get(`/3/tv/${id}?api_key=${credentials.tmdb.api_key}`);
-    res.send(data);
-  } catch (err) {
-    next(err);
-  }
-});
+app.use('/', routes);
 
 app.listen(3000);
